requirements.txt
ArticutAPI==1.1.5
cn2an==0.5.11
python-dotenv==0.18.0
pandas==1.3.0
